Fix bugs in Readign of geometrical data

+++ src/main/clm_varsur.F90
@@ -56,12 +56,12 @@ module clm_instur
   ! excess ice tiling mask
   integer, pointer :: exice_tile_mask(:) 
   
-  integer, pointer :: a_tile1(:)
+  real(r8), pointer :: a_tile1(:)
   
-  integer, pointer :: a_tile2(:)
+  real(r8), pointer :: a_tile2(:)
   
-  integer, pointer :: tile_dist(:)
+  real(r8), pointer :: tile_dist(:)
   
-  integer, pointer :: tile_ctl(:)
+  real(r8), pointer :: tile_ctl(:)
 
 end module clm_instur
